+++ previous-versions/accounting/src/codataccounting/models/operations/list_invoices.py
@@ -5,7 +5,7 @@ import dataclasses
 import requests as requests_http
 from ..shared import errormessage as shared_errormessage
 from ..shared import invoices as shared_invoices
-from typing import Optional
+from typing import Callable, Optional
 
 
 
@@ -27,6 +27,7 @@ class ListInvoicesRequest:
 
 @dataclasses.dataclass
 class ListInvoicesResponse:
+    next: Callable[[], Optional[ListInvoicesResponse]] = dataclasses.field()
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)