+++ previous-versions/accounting/src/codataccounting/models/operations/list_sales_orders.py
@@ -5,7 +5,7 @@ import dataclasses
 import requests as requests_http
 from ..shared import errormessage as shared_errormessage
 from ..shared import salesorders as shared_salesorders
-from typing import Optional
+from typing import Callable, Optional
 
 
 
@@ -27,6 +27,7 @@ class ListSalesOrdersRequest:
 
 @dataclasses.dataclass
 class ListSalesOrdersResponse:
+    next: Callable[[], Optional[ListSalesOrdersResponse]] = dataclasses.field()
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)