+++ previous-versions/accounting/src/codataccounting/models/operations/list_transfers.py
@@ -5,7 +5,7 @@ import dataclasses
 import requests as requests_http
 from ..shared import errormessage as shared_errormessage
 from ..shared import transfers as shared_transfers
-from typing import Optional
+from typing import Callable, Optional
 
 
 
@@ -28,6 +28,7 @@ class ListTransfersRequest:
 
 @dataclasses.dataclass
 class ListTransfersResponse:
+    next: Callable[[], Optional[ListTransfersResponse]] = dataclasses.field()
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)