+++ previous-versions/accounting/RELEASES.md
@@ -619,3 +619,13 @@ Based on:
 - [python v0.1.1] previous-versions/accounting
 ### Releases
 - [PyPI v0.1.1] https://pypi.org/project/codat-accounting/0.1.1 - previous-versions/accounting
+
+## 2023-09-12 00:10:47
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.82.5 (2.108.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.2] previous-versions/accounting
+### Releases
+- [PyPI v0.1.2] https://pypi.org/project/codat-accounting/0.1.2 - previous-versions/accounting